Changed code to use double goals instead of integer. Changing code to use any type of state when processing the match instead of just GDL state.

+++ src/main/java/csironi/ggp/course/experiments/tournaments/independentprocesses/IndependentSingleMatchRunner.java
@@ -441,7 +441,7 @@ public class IndependentSingleMatchRunner {
 
 			// Save the goals in the "/scores" file for the tournament.
 
-			List<Integer> goals;
+			List<Double> goals;
 			try {
 				goals = server.getGoals();
