+++ src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/treestructure/propnet/slowsequential/PnSlowSequentialTreeNodeFactory.java
@@ -30,7 +30,7 @@ public class PnSlowSequentialTreeNodeFactory implements PnTreeNodeFactory {
 	@Override
 	public MctsNode createNewNode(CompactMachineState state) {
 
-		int goals[] = null;
+		double goals[] = null;
 		boolean terminal = false;
 		List<List<CompactMove>> allLegalMoves = null;
