+++ src/main/java/org/ggp/base/util/statemachine/cache/NoSyncRefactoredSeparateInternalPropnetCachedStateMachine.java
@@ -35,13 +35,13 @@ public final class NoSyncRefactoredSeparateInternalPropnetCachedStateMachine ext
 	private final RefactoredTtlCache<CompactMachineState, PropnetMachineStateEntry> internalStateTtlCache;
 
 	private final class PropnetMachineStateEntry{
-		public Map<CompactRole, List<Integer>> goals;
+		public Map<CompactRole, List<Double>> goals;
 		public Map<CompactRole, List<CompactMove>> moves;
 		public Map<List<CompactMove>, CompactMachineState> nexts;
 		public Boolean terminal;
 
 		public PropnetMachineStateEntry(){
-			goals = new HashMap<CompactRole, List<Integer>>();
+			goals = new HashMap<CompactRole, List<Double>>();
 			moves = new HashMap<CompactRole, List<CompactMove>>();
 			nexts = new HashMap<List<CompactMove>, CompactMachineState>();
 			terminal = null;
@@ -80,7 +80,7 @@ public final class NoSyncRefactoredSeparateInternalPropnetCachedStateMachine ext
 	}
 
 	@Override
-	public List<Integer> getAllGoalsForOneRole(ExplicitMachineState state, ExplicitRole role){
+	public List<Double> getAllGoalsForOneRole(ExplicitMachineState state, ExplicitRole role){
 
 		//System.out.println("PN: Wrong call of cache (goals)!");
 
@@ -88,12 +88,12 @@ public final class NoSyncRefactoredSeparateInternalPropnetCachedStateMachine ext
 	}
 
 	@Override
-	public List<Integer> getAllGoalsForOneRole(CompactMachineState state, CompactRole role){
+	public List<Double> getAllGoalsForOneRole(CompactMachineState state, CompactRole role){
 		PropnetMachineStateEntry entry = getPropnetEntry(state);
 
 		//System.out.println("PN: Looking for goals in the cache!");
 
-		List<Integer> goals = entry.goals.get(role);
+		List<Double> goals = entry.goals.get(role);
 
 		if (goals == null){
